modify type of sender_len to int32

+++ include/evmc/evmc.h
@@ -112,13 +112,13 @@ struct evmc_message
     /** The destination of the message. */
     evmc_address destination;
     const uint8_t* destination_ptr;
-    size_t destination_len;
+    int32_t destination_len;
 
 
     /** The sender of the message. */
     evmc_address sender;
     const uint8_t* sender_ptr;
-    size_t sender_len;
+    int32_t sender_len;
 
     /**
      * The message input data.
